+++ src/Lights/SingleChannel.h
@@ -38,7 +38,7 @@ class SingleChannel: public Element {
 
         void SetIntensity(float targetI, float fadeTime);
     
-        void FollowSignal(int Amin, int Amax, signalState signal, int freq, int phase, float randomComponent);
+        void FollowSignal(int Amin, int Amax, signalState signal, float freq, int phase, float randomComponent);
 
 };
